% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{detectHubs}
\alias{detectHubs}
\title{Detect Haywire Hubs}
\usage{
detectHubs(object, gene_of_interest, tau, epsilon)
}
\arguments{
\item{delta_p_list}{List of Delta P vectors (output from calcRewiring).}
}
\value{
Data frame of genes ranked by \eqn{\sum |\Delta P|}.
}
\description{
Identifies master regulators by calculating Differential Degree Centrality (\eqn{\Delta k}).
}
