% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{dyprop-validity}
\alias{dyprop-validity}
\title{Validity Check for dyprop}
\arguments{
\item{object}{A dyprop object.}
}
\value{
TRUE if valid, otherwise a character string describing the error.
}
\description{
Validity Check for dyprop
}
