% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_transform.R
\name{prepareComposition}
\alias{prepareComposition}
\title{Prepare Compositional Data}
\usage{
prepareComposition(
  counts,
  pseudotime,
  design = matrix(nrow = 0, ncol = 0),
  k = 5,
  impute_zeros = TRUE
)
}
\arguments{
\item{counts}{Matrix. Raw counts (Samples x Genes).}

\item{pseudotime}{Numeric vector.}

\item{design}{Matrix. Design matrix for validation (optional).}

\item{k}{Integer. Number of neighbors for kNN. Default 5.}
}
\value{
A \code{dyprop} object ready for scanning.
}
\description{
A wrapper function to perform the pre-processing pipeline:
1. Technical Dropout Correction (kNN pooling)
2. Biological Zero Imputation (Bayesian Replacement)
3. CLR Transformation
}
